code optimization and component refactoring and organisation
