+++ platform/src/pages/Home/index.jsx
@@ -1,5 +1,5 @@
 import withAuth from '@/core/utils/protectedRoute';
-import React, { useState, useRef, useEffect } from 'react';
+import React, { useState, useEffect } from 'react';
 import Layout from '@/components/Layout';
 import CheckIcon from '@/icons/tickIcon';
 import Link from 'next/link';
@@ -7,213 +7,12 @@ import Button from '@/components/Button';
 import Image from 'next/image';
 import AnalyticsImage from '@/images/Home/analyticsImage.png';
 import PlayIcon from '@/images/Home/playIcon.svg';
-import CloseIcon from '@/icons/close_icon';
-import Spinner from '@/components/Spinner';
 import AnalyticsVideo from '../../../public/videos/analytics.mp4';
 import { useSelector, useDispatch } from 'react-redux';
-import {
-  startTask,
-  completeTask,
-  updateTitle,
-  updateVideoProgress,
-} from '@/lib/store/services/checklists/CheckList';
-
-const StepProgress = ({ step, totalSteps }) => {
-  const radius = 45;
-  const stroke = 10;
-  const normalizedRadius = radius - stroke * 2;
-  const circumference = normalizedRadius * 2 * Math.PI;
-  const strokeDashoffset = circumference - (step / totalSteps) * circumference;
-
-  return (
-    <div className='flex items-center justify-end'>
-      <svg height={radius * 2} width={radius * 2}>
-        <defs>
-          <mask id='round'>
-            <rect width='100%' height='100%' fill='#fff' />
-            <circle cx={radius} cy={radius} r={normalizedRadius - stroke / 2} fill='#000' />
-          </mask>
-        </defs>
-        <circle
-          stroke='#E1E7EC'
-          fill='transparent'
-          strokeWidth={stroke}
-          r={normalizedRadius}
-          cx={radius}
-          cy={radius}
-        />
-        <circle
-          stroke='#145FFF'
-          fill='transparent'
-          strokeWidth={stroke}
-          strokeDasharray={circumference + ' ' + circumference}
-          style={{ strokeDashoffset, transform: 'rotate(-90deg)', transformOrigin: '50% 50%' }}
-          r={normalizedRadius}
-          cx={radius}
-          cy={radius}
-          mask='url(#round)'
-        />
-        <text x='50%' y='50%' textAnchor='middle' fill='#536A87' fontSize='18px' dy='.4em'>
-          {step}/{totalSteps}
-        </text>
-      </svg>
-    </div>
-  );
-};
-
-const CustomModal = ({ open, setOpen, videoUrl, checklistData }) => {
-  const dispatch = useDispatch();
-  const modalRef = useRef(null);
-  const backdropRef = useRef(null);
-  const videoRef = useRef(null);
-  const [loading, setLoading] = useState(true);
-
-  const handleClickOutside = (event) => {
-    if (backdropRef.current && !backdropRef.current.contains(event.target)) {
-      setOpen(false);
-      dispatch(updateVideoProgress({ id: 1, videoProgress: videoRef.current.currentTime }));
-    }
-  };
-
-  useEffect(() => {
-    if (open) {
-      modalRef.current.focus();
-      document.addEventListener('mousedown', handleClickOutside);
-      const videoTime = checklistData.find(
-        (card) => card.title === 'analytics_video',
-      )?.videoProgress;
-      if (videoTime) {
-        videoRef.current.currentTime = videoTime;
-      }
-    } else {
-      document.removeEventListener('mousedown', handleClickOutside);
-      if (videoRef.current) {
-        dispatch(updateVideoProgress({ id: 1, videoProgress: videoRef.current.currentTime }));
-      }
-    }
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, [open]);
-
-  const handleVideoLoad = () => {
-    setLoading(false);
-  };
-
-  const handleVideoPause = () => {
-    dispatch(updateVideoProgress({ id: 1, videoProgress: videoRef.current.currentTime }));
-  };
-
-  const handleVideoEnd = () => {
-    dispatch(updateVideoProgress({ id: 1, videoProgress: 0 }));
-    dispatch(completeTask(1));
-  };
-
-  return (
-    open && (
-      <div
-        ref={modalRef}
-        id='custom-modal'
-        tabIndex='-1'
-        className={`fixed top-0 left-0 right-0 bottom-0 z-50 flex items-center justify-center transform transition-transform duration-300 ${
-          open ? 'opacity-100 translate-y-0' : 'opacity-0 -translate-y-2'
-        } bg-gray-800 bg-opacity-50`}>
-        <div
-          ref={backdropRef}
-          className='relative w-full h-[450px] max-w-2xl max-h-full bg-white rounded-xl shadow dark:bg-gray-700 p-1 mx-4 md:mx-0 animate-slide-in'>
-          <h1 className='text-sm md:text-xl w-full text-center font-medium text-gray-900 p-3'>
-            Introducing AirQo Analytics
-          </h1>
-          <button
-            type='button'
-            onClick={() => {
-              setOpen(!open);
-              dispatch(updateVideoProgress({ id: 1, videoProgress: videoRef.current.currentTime }));
-            }}
-            className='absolute top-0 right-0 md:-top-[25px] md:-right-[24px] m-2 text-gray-400 bg-blue-600 hover:bg-blue-900 hover:text-gray-900 rounded-full text-sm w-8 h-8 inline-flex justify-center items-center dark:hover:bg-gray-600 dark:hover:text-white'
-            data-modal-hide='custom-modal'>
-            <CloseIcon fill='#FFFFFF' />
-            <span className='sr-only'>Close modal</span>
-          </button>
-          <div className='w-full h-full max-h-[384px] relative'>
-            {loading && (
-              <div className='absolute top-0 left-0 w-full h-full flex justify-center items-center'>
-                <Spinner />
-              </div>
-            )}
-            <video
-              ref={videoRef}
-              onCanPlayThrough={handleVideoLoad}
-              onPause={handleVideoPause}
-              onEnded={handleVideoEnd}
-              className={`w-full h-full ${loading ? 'hidden' : 'block'}`}
-              src={videoUrl}
-              controls
-              autoPlay
-            />
-          </div>
-        </div>
-      </div>
-    )
-  );
-};
-
-const HomeSkeleton = () => {
-  return (
-    <div className='px-3 lg:px-16 py-3 space-y-5'>
-      <div className='w-full mb-4 md:mb-10 animate-pulse'>
-        <h1 className='text-[32px] leading-10 font-medium bg-gray-300 h-10 w-3/4'></h1>
-      </div>
-
-      <div className='w-full flex justify-between items-center'>
-        <div className='w-full flex flex-col items-start'>
-          <h1 className='text-2xl font-medium text-gray-900 bg-gray-300 h-7 w-1/2'></h1>
-          <p className='text-sm font-normal text-gray-500 bg-gray-300 h-5 w-3/4'></p>
-        </div>
-        <div className='w-full'>
-          <div className='h-5 bg-gray-300 w-1/4'></div>
-        </div>
-      </div>
-
-      <div className='w-full grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-5'>
-        {[1, 2, 3, 4].map((index) => (
-          <div
-            key={index}
-            className='w-full h-[250px] flex flex-col justify-between items-start border-[0.5px] rounded-xl border-grey-150 py-5 px-3 space-y-5 focus:outline-blue-600 focus:ring-2 focus:shadow-lg focus:border-blue-600 animate-pulse'
-            tabIndex={0}>
-            <div className='w-full bg-gray-300 h-14 flex justify-center items-center rounded-full'></div>
-            <div className='w-full text-base font-normal bg-gray-300 h-5'></div>
-            <div className={`w-full text-sm flex justify-between font-normal bg-gray-300 h-5`}>
-              <span className='text-blue-600 bg-gray-300 h-4 w-1/4'></span>
-              <span className='text-black-900 bg-gray-300 h-4 w-1/4'></span>
-            </div>
-          </div>
-        ))}
-      </div>
-
-      <div className='w-full grid grid-cols-1 md:grid-cols-2 gap-3 items-center border-[0.5px] rounded-xl border-grey-150 p-3'>
-        <div className='flex flex-col justify-start p-8 animate-pulse'>
-          <h1 className='text-black-900 text-2xl font-medium bg-gray-300 h-7 w-3/4'></h1>
-          <p className='text-lg font-normal text-black-900 mt-2 bg-gray-300 h-5 w-3/4'></p>
-          <div className='mt-4 flex items-center space-x-8'>
-            <div className='bg-gray-300 text-white rounded-lg w-32 h-12'></div>
-            <div className='text-gray-300 text-sm font-normal mt-2 cursor-pointer bg-gray-300 h-5 w-1/4'></div>
-          </div>
-        </div>
-        <div
-          className='rounded-md p-9 relative bg-gray-300'
-          style={{
-            background: '#145DFF08',
-          }}>
-          <div className='absolute z-50 inset-0 flex items-center justify-center cursor-pointer'>
-            <div className='w-8 h-8 bg-gray-300 rounded-full'></div>
-          </div>
-          <div className='bg-gray-300 h-48 w-72'></div>
-        </div>
-      </div>
-    </div>
-  );
-};
+import { startTask, updateTitle } from '@/lib/store/services/checklists/CheckList';
+import HomeSkeleton from '@/components/skeletons/HomeSkeleton';
+import CustomModal from '@/components/Modal/videoModals/CustomModal';
+import StepProgress from '@/components/steppers/CircularStepper';
 
 const Home = () => {
   const dispatch = useDispatch();
